# backend/dslearn/api-docs.yml
openapi: 3.0.1
info:
  title: DS-Learn API
  description: API documentation for DS-Learn - Data Structures and Algorithms Learning Platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/exercises:
    get:
      tags:
        - Exercises
      summary: Get exercises paginated
      description: Retrieve exercises with pagination and optional filters
      operationId: getExercisesPaginated
      parameters:
        - name: difficulty
          in: query
          description: Filter by difficulty level
          required: false
          schema:
            type: string
        - name: dataStructure
          in: query
          description: Filter by data structure
          required: false
          schema:
            type: string
        - name: company
          in: query
          description: Filter by company
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sortBy
          in: query
          description: Sort by field
          required: false
          schema:
            type: string
            default: id
        - name: sortDirection
          in: query
          description: Sort direction (asc/desc)
          required: false
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Exercises retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedResponseDTOExerciseDTO'
components:
  schemas:
...
    ExerciseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        url:
          type: string
        difficulty:
          type: string
        relatedTopics:
          type: string
        companies:
          type: string
    PagedResponseDTOExerciseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseDTO'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
  securitySchemes:
    bearer-token:
      type: http
      description: Enter your Google ID token
      scheme: bearer
      bearerFormat: JWT
    token-header:
      type: apiKey
      description: Google ID token in header
      name: Token
      in: header